Fix float conversion in test

+++ tests/neighbors/test_faiss.py
@@ -54,8 +54,8 @@ def test_different_instantiations(single_source, source_target):
 )
 def test_torch_gpu(hubness, source_target):
     k = 3
-    source = torch.tensor(source_target[0]).cuda()
-    target = torch.tensor(source_target[1]).cuda()
+    source = torch.tensor(source_target[0]).to(torch.float32).cuda()
+    target = torch.tensor(source_target[1]).to(torch.float32).cuda()
     device_type = source.device.type
     nn_inst = Faiss(metric="l2", index_key="Flat", use_gpu=True)
     kiez_inst = Kiez(n_candidates=5, algorithm=nn_inst, hubness=hubness)
